requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-material==0.8.0
python-decouple==3.0
static3==0.7.0
gunicorn==19.4.1
psycopg2==2.6.1
djangorestframework==3.5.3